backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.48
aiosqlite==0.20.0
pydantic-settings==2.7.1
python-multipart==0.0.20
pymupdf==1.25.3
anthropic==0.43.0
google-genai>=1.69.0
elevenlabs==1.50.0
httpx==0.28.1
websockets==14.1
aiofiles==24.1.0
lameenc>=1.7.0
google-cloud-texttospeech>=2.21.0
